% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat01.R
\docType{data}
\name{dat01}
\alias{dat01}
\title{Dichotomous Response Dataset for IRT Calibration}
\format{
A data frame with rows representing examinees and 20 columns
representing items (ITEM001 to ITEM023). Responses are dichotomous
where 1 indicates a correct response and 0 indicates an incorrect response.
}
\usage{
data(dat01)
}
\description{
A sample dataset containing binary examination responses used to demonstrate
2PL model parameter estimation.
}
\keyword{datasets}
