\name{dat01}
\alias{dat01}
\docType{data}
\title{
dat01/dichotomous
}
\description{
 A concise (1-5 lines) description of the dataset.
}
\usage{data("dat01")}
\format{
  A data frame with 180 observations on the following 20 variables.
  \describe{
    \item{\code{ITEM001}}{a numeric vector}
    \item{\code{ITEM002}}{a numeric vector}
    \item{\code{ITEM003}}{a numeric vector}
    \item{\code{ITEM005}}{a numeric vector}
    \item{\code{ITEM006}}{a numeric vector}
    \item{\code{ITEM007}}{a numeric vector}
    \item{\code{ITEM008}}{a numeric vector}
    \item{\code{ITEM009}}{a numeric vector}
    \item{\code{ITEM010}}{a numeric vector}
    \item{\code{ITEM011}}{a numeric vector}
    \item{\code{ITEM012}}{a numeric vector}
    \item{\code{ITEM014}}{a numeric vector}
    \item{\code{ITEM015}}{a numeric vector}
    \item{\code{ITEM016}}{a numeric vector}
    \item{\code{ITEM017}}{a numeric vector}
    \item{\code{ITEM018}}{a numeric vector}
    \item{\code{ITEM019}}{a numeric vector}
    \item{\code{ITEM020}}{a numeric vector}
    \item{\code{ITEM021}}{a numeric vector}
    \item{\code{ITEM023}}{a numeric vector}
      }
}
\details{
 A matrix or data.frame consists of dichotomous data (1 for correct and 0 for wrong response)
}
\source{
reference to a publication or URL from which the data were obtained
}
\references{
possibly secondary sources and usages
}
\examples{
data(dat01)
## maybe str(dat01) ; plot(dat01) ...
}
\keyword{datasets}
