% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtbem2pl.R
\name{irtbem2pl}
\alias{irtbem2pl}
\title{Calibrating 2PL model via (a,b), proposed by Birnbaum
(1968). Adjustment with Marginal Bayesian Modal Item Parameter Estimation.}
\usage{
irtbem2pl(
  data,
  PriorA = c(0, 0.25),
  PriorB = c(0, 4),
  InitialA = 1,
  InitialB = 0,
  Tol = 0.01,
  max.ECycle = 100L,
  max.MCycle = 3L,
  n.decimal = 5L,
  n.Quadpts = 50L,
  Theta.lim = c(-4, 4),
  Missing = -9,
  ParConstraint = F,
  BiasSE = F,
  D = 1
)
}
\arguments{
\item{data}{A matrix or data.frame consists of dichotomous data
(1 for correct and 0 for wrong response)}

\item{PriorA}{The user specified logarithmic normal distribution prior for item
discrimation (a) parameters in the 2PL models.
A numeric with two hyperparameters
mean and variance of logarithmic normal distribution for all a parameters.
By default, PriorA=c(0,0.25), which means a log normal prior of mean=0 and
variance=0.25 will be used for all item discrimation parameters.}

\item{PriorB}{The user specified normal distribution prior for item difficulty (b)
parameters in the 2PL models.
A numeric with two hyperparameters mean and variance of normal distribution for
all b parameters. By default, PriorB=c(0,4), which means a normal prior of
mean=0 and variance=4 will be used for all item difficulty parameters.}

\item{InitialA}{The user specified starting values for item discrimation (a)
parameters in the 2PL model. A 1 (default), single number (numeric), refers to set t
his number to be the starting values of a for all items}

\item{InitialB}{The user specified starting values for item difficulty (b)
parameters in the 2PL model. A 0 (default), single number (numeric),
refers to set this number to be the starting values of b for all items.}

\item{Tol}{A single number (numeric), refers to convergence threshold
for E-step cycles; defaults are 0.01.}

\item{max.ECycle}{A single integer, refers to maximum number of E-step cycles;
defaults are 100L.}

\item{max.MCycle}{A single integer, refers to maximum number of M-step
cycles; defaults are 3L.}

\item{n.decimal}{A single integer, refers to number of decimal places when o
utputs results. defaults are 5L.}

\item{n.Quadpts}{A single integer, refers to number of quadrature points
per dimension (must be larger than 5); defaults are 50L.}

\item{Theta.lim}{A numeric with two number, refers to the range of
integration grid for each dimension; default is c(-4, 4).}

\item{Missing}{A single number (numeric) to indicate which elements are missing;
default is -9. The Missing cannot be 0 or 1.}

\item{ParConstraint}{A logical value to indicate whether estimates
parametes in a reasonable range; default is FALSE.}

\item{BiasSE}{A logical value to determine whether directly estimating
SEs from inversed Hession matrix rather than USEM method, default is FALSE}

\item{D}{the scaling constant 1 normal or 1.702 log.}
}
\value{
This function will return a list includes following:
\itemize{
\item \verb{Est.ItemPars (a,b):} A dataframe consists of the estimates of a, b and c
and corresponding estimated standard errors.
\item \verb{Est.Theta (ability scale:)} A dataframe consists of the estimates of theta
and corresponding estimated standard errors (EAP method).
\item \verb{Loglikelihood:} The loglikelihood.
\item \verb{Iteration The:} number of iterations.
\item \verb{EM.Map:} The parameter estimation history of iterations.
\item \verb{fits.test:} The model fits information includes G2 test, AIC, BIC and RMSEA.
\item \verb{Elapsed.time:} The running time of the program.
\item \verb{InitialValues:} The initial values of item parameters.
}
}
\description{
This function can estimate the item parameters of the 2PL model
via  Bayesian statistical methodology include Cornfield
(1969),  Finetti (1974), Edwards, Lindman, and Savage (1963), Lindley
(1970a, 1970b, 1971), and Novick and Jackson (1974). Examples of the use
of these methods in educational settings can be found in Novick and Jackson
(1974), Novick, Jackson, Thayer, and Cole (1972), and Rubin (1980). Lord
(1986) compared maximum likelihood and Bayesian estimation methods in
IRT.
}
\details{
Two parameter logistic (2PL) model proposed by Birnbaum (1968):

P(x=1∣θ,a,b)=(1+exp(−D∗a∗(θ−b)))

where x=1 is the correct response, theta is examinne's ability;
a, b are the item discrimination, and difficulty parameter,
respectively; D is the scaling constant 1 normal.

The implementation of expected a posteriori (EAP) ability estimation
has communalities with the E-step of both marginal maximum likelihood /
expectation-maximization (MMLE/EM) and Bayesian modal estimation /
expectation-maximization (MBE/EM) estimation of item parameters.
}
\examples{
data(dat01)
library(irtbem2pl)
 mod_2PL<-irtbem2pl(dat01)

 mod_2PL$Est.ItemPars       #show item estimates
 mod_2PL$Est.Theta          #show ability estimates

}
\references{
\emph{Item Response Theory Parameter Estimation Techniques}
Second Edition, Revised and Expanded Frank B. Baker University ofWisconsin
Madison, Wisconsin, U.S.A. Seock-Ho Kim The University ofGeorgia
Athens, Georgia, u.S.A.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an
examinee's ability. In F. M. Lord & M. R. Novick (Eds.), Statistical theories
of mental test scores (pp. 395-479). MA: Adison-Wesley.

Guo, S., & Zheng, C. (2019). The Bayesian Expectation-Maximization-Maximization
for the 3PLM. Frontiers in Psychology, 10(1175), 1-11. doi:10.3389/fpsyg.2019.01175

Zheng, C., Meng, X., Guo, S., & Liu, Z. (2018). Expectation-Maximization-Maximization:
A feasible MLE algorithm for the three-parameter logistic model based on a mixture
modeling reformulation. Frontiers in Psychology, 8(2302), 1-10. doi:10.3389/fpsyg.2017.02302
}
\seealso{
\code{\link{Input.Checking}}, \code{\link{BEMM.3PL}},
\code{\link{BEMM.1PLG}}, \code{\link{Prob.model}}
}
\author{
Juan Luis Legorreta Torres \email{jlegorreta2002@yahoo.com.mx}
}
