% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtchek2pl.R
\name{irtchek2pl}
\alias{irtchek2pl}
\title{Checking user speciflied input variables}
\usage{
irtchek2pl(
  Model = "2PL",
  data,
  PriorA = c(0, 0.25),
  PriorB = c(0, 4),
  InitialA = 1,
  InitialB = 0,
  Tol = 0.01,
  max.ECycle = 50L,
  max.MCycle = 10L,
  n.decimal = 5L,
  n.Quadpts = 50L,
  Theta.lim = c(-4, 4),
  Missing = -9,
  ParConstraint = FALSE,
  BiasSE = FALSE,
  D = 1
)
}
\arguments{
\item{Model}{A character to declare the type of items to be modeled.
The parameter labels follow conventional use.
'2pl' - Two parameter logistic (3PL) mode

P(x=1|Theta,a,b)=1+exp(-D\emph{a}(Theta-b))
where x=1 is the correct response, theta is examinne's ability;
a, b are the item discrimination, and difficulty parameter,
respectively; D is the scaling constant 1 normal.}

\item{data}{A matrix or data.frame consists of dichotomous data
(1 for correct and 0 for wrong response), with missing data coded
as in Missing (by default, Missing=-9). Each row of data represents
a examinne' responses, and each column represents an}

\item{PriorA}{The user specified logarithmic normal distribution prior for item
discrimation (a) parameters in the 2PL models.
A numeric with two hyperparameters
mean and variance of logarithmic normal distribution for all a parameters.
By default, PriorA=c(0,0.25), which means a log normal prior of mean=0 and
variance=0.25 will be used for all item discrimation parameters.}

\item{PriorB}{The user specified normal distribution prior for item difficulty (b)
parameters in the 2PL models.
A numeric with two hyperparameters mean and variance of normal distribution for
all b parameters. By default, PriorB=c(0,4), which means a normal prior of
mean=0 and variance=4 will be used for all item difficulty parameters.}

\item{InitialA}{The user specified starting values for item discrimation (a)
parameters in the 2PL model. A 1 (default), single number (numeric), refers to set t
his number to be the starting values of a for all items}

\item{InitialB}{The user specified starting values for item difficulty (b)
parameters in the 2PL model. A 0 (default), single number (numeric),
refers to set this number to be the starting values of b for all items.}

\item{Tol}{A single number (numeric), refers to convergence threshold
for E-step cycles; defaults are 0.01.}

\item{max.ECycle}{A single integer, refers to maximum number of E-step cycles;
defaults are 100L.}

\item{max.MCycle}{A single integer, refers to maximum number of M-step
cycles; defaults are 3L.}

\item{n.decimal}{A single integer, refers to number of decimal places when o
utputs results. defaults are 5L.}

\item{n.Quadpts}{A single integer, refers to number of quadrature points
per dimension (must be larger than 5); defaults are 50L.}

\item{Theta.lim}{A numeric with two number, refers to the range of
integration grid for each dimension; default is c(-4, 4).}

\item{Missing}{A single number (numeric) to indicate which elements are missing;
default is -9. The Missing cannot be 0 or 1.}

\item{ParConstraint}{A logical value to indicate whether estimates
parametes in a reasonable range; default is FALSE.}

\item{BiasSE}{A logical value to determine whether directly estimating
SEs from inversed Hession matrix rather than USEM method, default is FALSE}

\item{D}{the scaling constant 1 normal or 1.702 log.}
}
\description{
Based on the given model, checking whether user
speciflied input variables are correct. If the input variables
are acceptable, this function will format them and then return them
as a list. Otherwise, this function will return a error message to
indicate which variables are unacceptable.
}
\examples{
data(dat01)
library(irtbem2pl)
 Checking_2PL<-irtchek2pl(Model = "2PL", dat01, D = 1)
 #STOP: Some elements in data are not 1, 0 or Missing
       #PriorA[2] is the variance, and it must bigger than 0!
      # PriorB must have two input values unless PriorB=NA
      #The min of Tol must bigger than 0
      # Theta.lim[1] must bigger than Theta.lim[2] etc ....

}
\references{
\emph{Item Response Theory Parameter Estimation Techniques}
Second Edition, Revised and Expanded Frank B. Baker University ofWisconsin
Madison, Wisconsin, U.S.A. Seock-Ho Kim The University ofGeorgia
Athens, Georgia, u.S.A.
}
\author{
Juan Luis Legorreta Torres \email{jlegorreta2002@yahoo.com.mx}
}
