% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodInfo2pl.R
\name{LikelihoodInfo2pl}
\alias{LikelihoodInfo2pl}
\title{Compute Likelihood Information for the 2PL IRT Model}
\usage{
LikelihoodInfo2pl(
  data.simple,
  CountNum,
  Model,
  Par.est0,
  n.Quadpts,
  node.Quadpts,
  weight.Quadpts,
  D
)
}
\arguments{
\item{data.simple}{A numeric matrix representing the simplified unique
response patterns.}

\item{CountNum}{A numeric vector containing the frequencies of each unique
response pattern.}

\item{Model}{A character string declaring the type of IRT model to be evaluated
(e.g., "2PL").}

\item{Par.est0}{A list containing current item parameter estimates (A and B).}

\item{n.Quadpts}{A single integer indicating the number of quadrature points.}

\item{node.Quadpts}{A numeric vector containing the nodes of the quadrature grid.}

\item{weight.Quadpts}{A numeric vector containing the weights of the quadrature grid.}

\item{D}{A numeric scaling constant (typically 1 or 1.702).}
}
\value{
A list containing the following statistical components:
\itemize{
\item \code{LH} The log-likelihood value of the model given the data.
\item \code{Deviance} The model deviance statistic calculated as -2 * LH.
\item \code{f} A numeric matrix of expected frequencies per quadrature point.
\item \code{fz} A numeric matrix of expected correct response frequencies based on P*.
\item \code{rz} A numeric matrix of expected correct joint frequencies.
}
}
\description{
' Calculates the log-likelihood, deviance, and expected frequencies (f, fz, rz)
required for the E-step of the Expectation-Maximization (EM) algorithm
under a two-parameter logistic (2PL) model.
}
