% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodInfo2pl.R
\name{LikelihoodInfo2pl}
\alias{LikelihoodInfo2pl}
\title{Maximum Likelihood Estimation (MLE)}
\usage{
LikelihoodInfo2pl(
  data.simple,
  CountNum,
  Model,
  Par.est0,
  n.Quadpts,
  node.Quadpts,
  weight.Quadpts,
  D
)
}
\arguments{
\item{CountNum}{..}

\item{Model}{Two parameter logistic (2PL) model}

\item{Par.est0}{A list that consists of item parameters for each item
based on the given model. For 2PL model, list(A, B)}

\item{n.Quadpts}{A single integer, refers to number of quadrature points
per dimension}

\item{node.Quadpts}{..}

\item{weight.Quadpts}{..}

\item{D}{the scaling constant.  By default 1 normal or 1.702 log.}

\item{data.siple}{Data frame con la base}
}
\description{
"Likelihood" is a statistical concept that measures how likely
it is to observe a data outcome given a set of model parameters.
Maximum-likelihood fitting of univariate distributions,
allowing parameters to be held fixed if desired
}
\seealso{
\code{\link{Input.Checking}}, \code{\link{BEMM.3PL}},
\code{\link{BEMM.1PLG}}, \code{\link{Prob.model}}
}
